Camera Animations, Optimised Levels
Camera Animations are added when player spawns, and when player is dead. Levels are optimised and all functionalities should work on all levels.

+++ GLD-MSD62A-Teamwork-Game/Assets/Scripts/ItemScriptableObject.cs
@@ -7,7 +7,7 @@ public class ItemScriptableObject : ScriptableObject
 {
     public string title;
     public Sprite icon;
-    public int increaseValue;
+    public float increaseValue;
     public int cost;
     public Type type; //stores either health or mana (not both at the same time)
